New CD pipeline (#91)
* Update event types and payload in CI workflow

The changes include the update of event types from specific to a common 'argocd' and 'argocd-manual'. Now, the 'commit_msg' in the payload includes branch references and the triggering actor for manual deployment. Additionally, 'environment' and 'ref' fields have been added to the payload.

* Refactor Github Actions workflows for CI/CD separation

Separate the continuous integration and continuous deployment steps in their respective Github Actions workflows for cleaner code structure and maintenance. This allows for granular control and easy update of each step in the pipeline.